invites

+++ Vineyard/Vineyard/Resolution.swift
@@ -32,11 +32,11 @@ class Resolution: Identifiable, Codable {
     let id: String
     var title: String
     var description: String
-    var defaultQuantity: Int? = nil
+    var defaultQuantity: Float? = nil
     var defaultFrequency: Frequency
     var diffLevel: Difficulty
     
-    init(title: String, description: String, quantity: Int? = nil, frequency: Frequency, diffLevel: Difficulty) {
+    init(title: String, description: String, quantity: Float? = nil, frequency: Frequency, diffLevel: Difficulty) {
         self.id = UUID().uuidString
         self.title = title
         self.description = description
